More work on the Load Structure gui and added the Logitow Hub gui.

+++ src/main/java/com/logitow/logimine/blocks/BlockKey.java
@@ -3,7 +3,7 @@ package com.logitow.logimine.blocks;
 import com.logitow.bridge.build.Structure;
 import com.logitow.logimine.LogiMine;
 import com.logitow.logimine.client.gui.DeviceManagerGui;
-import com.logitow.logimine.client.gui.ManagerChoiceGui;
+import com.logitow.logimine.client.gui.HubGui;
 import com.logitow.logimine.items.ModItems;
 import com.logitow.logimine.tiles.TileEntityBlockKey;
 import net.minecraft.block.ITileEntityProvider;
@@ -80,9 +80,9 @@ public class BlockKey extends BlockBase implements ITileEntityProvider {
                             //Rotating the structures.
                             return blockKeyEntity.rotateStructure(player, facing);
                         } else if(world.isRemote) {
-                            Minecraft.getMinecraft().displayGuiScreen(new DeviceManagerGui());
+                            Minecraft.getMinecraft().displayGuiScreen(new HubGui());
                             //Assigning the block to the device manager dialog.
-                            ManagerChoiceGui.setSelectedKeyBlock(blockpos);
+                            HubGui.setSelectedKeyBlock(blockpos);
                             return true;
                         }
                     }
@@ -100,7 +100,7 @@ public class BlockKey extends BlockBase implements ITileEntityProvider {
         System.out.println("Destroyed key block at: " + pos);
         if(worldIn.isRemote) { //Client
             //Closing the gui.
-            if(DeviceManagerGui.instance != null && ManagerChoiceGui.instance.getSelectedKeyBlock() != null && ManagerChoiceGui.instance.getSelectedKeyBlock().getPos().equals(pos)) {
+            if(DeviceManagerGui.instance != null && HubGui.instance.getSelectedKeyBlock() != null && HubGui.instance.getSelectedKeyBlock().getPos().equals(pos)) {
                 Minecraft.getMinecraft().displayGuiScreen(null);
             }
         }
@@ -120,7 +120,9 @@ public class BlockKey extends BlockBase implements ITileEntityProvider {
                     if(!keyblock.getWorld().isRemote) {
                         //Deleting structures file.
                         if(keyblock.getAssignedStructure() != null) {
-                            Structure.removeFile(keyblock.getAssignedStructure());
+                            if(keyblock.getAssignedStructure().customName == null) {
+                                Structure.removeFile(keyblock.getAssignedStructure());
+                            }
                             keyblock.clearStructure();
                         }
                         keyblock.assignDevice(null,null);
